refactored oven handler

+++ bundles/org.openhab.binding.homeconnect/src/main/java/org/openhab/binding/homeconnect/internal/client/model/Event.java
@@ -43,7 +43,7 @@ public class Event {
     }
 
     public int getValueAsInt() {
-        return value != null ? Integer.valueOf(getValue()).intValue() : 0;
+        return value != null ? Float.valueOf(getValue()).intValue() : 0;
     }
 
     public String getUnit() {
